Create Poem
"A Poison Tree" is a poem written by William Blake, published in 1794 as part of his Songs of Experience collection. It describes the narrator's repressed feelings of anger towards an individual, emotions which eventually lead to murder. The poem explores themes of indignation, revenge, and more generally the fallen state of mankind.